Move code into dedicated files
